// backend/src/modules/repairs/enums/estado-reparacion.enum.ts
export enum EstadoReparacion {
  INGRESADO = 'INGRESADO',
  EN_COLA_DIAGNOSTICO = 'EN_COLA_DIAGNOSTICO',
  EN_DIAGNOSTICO = 'EN_DIAGNOSTICO',
  PRESUPUESTADO_ESPERANDO_OK = 'PRESUPUESTADO_ESPERANDO_OK',
  PRESUPUESTO_RECHAZADO = 'PRESUPUESTO_RECHAZADO',
  RESPALDO_DE_DATOS = 'RESPALDO_DE_DATOS',
  EN_REPARACION = 'EN_REPARACION',
  ESPERANDO_REPUESTO_LOCAL = 'ESPERANDO_REPUESTO_LOCAL',
  ESPERANDO_REPUESTO_IMPORTACION = 'ESPERANDO_REPUESTO_IMPORTACION',
  EN_PRUEBAS_CONTROL_CALIDAD = 'EN_PRUEBAS_CONTROL_CALIDAD',
  REPARADO_PENDIENTE_PAGO = 'REPARADO_PENDIENTE_PAGO',
  LISTO_PARA_RETIRAR = 'LISTO_PARA_RETIRAR',
  ENTREGADO_AL_CLIENTE = 'ENTREGADO_AL_CLIENTE',
  CERRADO_FACTURADO = 'CERRADO_FACTURADO',
  IRREPARABLE_PARA_RETIRAR = 'IRREPARABLE_PARA_RETIRAR',
  IRREPARABLE_ENTREGADO = 'IRREPARABLE_ENTREGADO',
  EN_GARANTIA_REINGRESO = 'EN_GARANTIA_REINGRESO',
  GARANTIA_ENTREGADO = 'GARANTIA_ENTREGADO',
  ABANDONADO_POR_CLIENTE = 'ABANDONADO_POR_CLIENTE',
  CANCELADO_POR_CLIENTE = 'CANCELADO_POR_CLIENTE',
}